+++ include/KeyFrame.h
@@ -124,7 +124,7 @@ public:
     long unsigned int mnId;
     const long unsigned int mnFrameId;
 
-    const double mTimeStamp;
+    const int64_t mTimeStamp;
 
     // Grid (to speed up feature matching)
     const int mnGridCols;
